move NodeRow into own component and migrate styles
